Rename win32_comspec_warn win32_comspec_load.
